:zap: :lipstick: Fetch Spotify track information only once
Previously there were at least two requests per track to the Spotify API: One for the search, and one for the metadata. We now cache the tracks we already have so that we do not need to hit the Spotify API twice. This improves our effiency and user experience, since there are no more FOUC occuring in the search results.