:hammer: Set state.currentTrack in provider-tracks
Right now it's provider-player's responsibility to put the currently playing track into the app state. This is a problem because we need a reference to the currently playing track, even if provider-player is not running because we're on the client side. provider-tracks, however, has a reference to the tracks as they are in Firebase, so it makes sense to set the currently playing track from there.