:bug: Fix invalid use of transactions
It should be noted that Firebase transactions abort _immediately_ when `undefined` is returned from the transaction function, even if it ran on stale data. We previously used `undefined` as a value to tell Firebase that we do not intend to change the data at that location. This is incorrect, however, as our observations could partially be based on stale data, and as such, we made wrong conclusions, causing Firebase to abort immediately and never giving us the actual data. Thus, we now simply return what we were given instead of `undefined`.